requirements.txt
numpy==1.26.4
transformers==4.47.1
datasets==3.2.0
torch
pytorch-lightning
torchmetrics
torchvision
scipy==1.13.1
pytest
huggingface_hub
ai2-olmo
accelerate
hydra-core
scikit-learn
matplotlib
nltk
tqdm
scikit-learn
openai==1.59.4
packaging
ninja
zstandard
simple-parsing
